# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API для работы с каталогом исполнителей, альбомов и песен.",
        "contact": {
            "email": "admin@musiccatalog.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Получить список всех исполнителей",
                "description": "Возвращает список всех исполнителей в базе данных",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "responses": {
                    "200": {
                        "description": "Успешный ответ с данными исполнителей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Artist"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Создать исполнителя",
                "description": "Создает исполнителя и возвращает его данные",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя исполнителя",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Исполнитель успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artist"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Обновить данные исполнителя",
                "operationId": "5a5632ea389f9a05336f8d035f5a44c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID исполнителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя исполнителя",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Исполнитель обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Исполнитель не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Удалить исполнителя",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID исполнителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Исполнитель удален"
                    },
                    "404": {
                        "description": "Исполнитель не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Artist": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID исполнителя",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Имя исполнителя",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}